+++ src/main/java/znu/visum/components/externals/domain/models/ExternalActor.java
@@ -1,0 +1,7 @@
+package znu.visum.components.externals.domain.models;
+
+public class ExternalActor extends ExternalPeople {
+  public ExternalActor(long id, String forename, String name) {
+    super(id, forename, name);
+  }
+}