+++ src/main/java/znu/visum/components/externals/domain/models/ExternalApi.java
@@ -1,0 +1,6 @@
+package znu.visum.components.externals.domain.models;
+
+/** Enum of the provider of Visum data. */
+public enum ExternalApi {
+  TMDB
+}