+++ src/main/java/znu/visum/components/externals/domain/models/ExternalDirector.java
@@ -1,0 +1,7 @@
+package znu.visum.components.externals.domain.models;
+
+public class ExternalDirector extends ExternalPeople {
+  public ExternalDirector(long id, String forename, String name) {
+    super(id, forename, name);
+  }
+}