+++ src/main/java/znu/visum/components/externals/tmdb/infrastructure/validators/TmdbGetConfigurationResponseBodyValidationHandler.java
@@ -1,0 +1,16 @@
+package znu.visum.components.externals.tmdb.infrastructure.validators;
+
+import znu.visum.components.externals.domain.models.ExternalApi;
+import znu.visum.components.externals.tmdb.infrastructure.models.TmdbGetConfigurationResponse;
+import znu.visum.core.validators.infrastructure.AbstractExternalResponseBodyValidationHandler;
+
+public class TmdbGetConfigurationResponseBodyValidationHandler
+    extends AbstractExternalResponseBodyValidationHandler<
+        TmdbGetConfigurationResponse, TmdbGetConfigurationResponseValidator> {
+  public TmdbGetConfigurationResponseBodyValidationHandler() {
+    super(
+        TmdbGetConfigurationResponse.class,
+        new TmdbGetConfigurationResponseValidator(),
+        ExternalApi.TMDB);
+  }
+}