+++ src/main/java/znu/visum/components/externals/tmdb/usecases/getmoviebyid/domain/GetTmdbMovieByIdService.java
@@ -1,0 +1,7 @@
+package znu.visum.components.externals.tmdb.usecases.getmoviebyid.domain;
+
+import znu.visum.components.externals.domain.models.ExternalMovie;
+
+public interface GetTmdbMovieByIdService {
+  ExternalMovie getTmdbMovieById(long id);
+}