+++ src/main/java/znu/visum/components/movies/usecases/getbyid/application/GetByIdMovieResponse.java
@@ -400,6 +400,10 @@ public class GetByIdMovieResponse {
       return runtime;
     }
 
+    public String getPosterUrl() {
+      return posterUrl;
+    }
+
     public static class Builder {
       private final ResponseMovieMetadata metadata;