+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.3.0
+
+* Add TMDb poster's URL in the `/tmdb/movies/search` and `/tmdb/movies/<id>` responses
+* Add `String getConfigurationBasePosterUrl();` to the `TmdbConnector` port and the implementation in the HTTP adapter
+* Add `SearchTmdbMoviesServiceImplUnitTest`
+* Rename `external` package to `externals`
+* Add [Caffeine](https://github.com/ben-manes/caffeine) in order to cache the TMDb (base) poster's URL with `@Cacheable`
+
 ## 2.2.1
 
 * Add pagination usecases IT